Project Dir refactor based on feature types. Feature package added: Notes, ScribblePad
