--- app/src/main/java/com/manujain/flashnotes/core/Utils.kt
@@ -1,14 +1,0 @@
-package com.manujain.flashnotes.core
-
-import com.manujain.flashnotes.domain.utils.NotesOrder
-import com.manujain.flashnotes.domain.utils.OrderType
-
-fun getNotesOrderFromStr(str: String, orderType: OrderType): NotesOrder {
-    // TODO This approach requires refactor
-    return if (str == "Title") {
-        NotesOrder.Title(orderType)
-    } else if (str == "Color")
-        NotesOrder.Color(orderType)
-    else
-        NotesOrder.Date(orderType)
-}
