--- app/src/main/java/com/manujain/flashnotes/domain/model/NoteState.kt
@@ -1,8 +1,0 @@
-package com.manujain.flashnotes.domain.model
-
-data class NoteState(
-    val title: String = "",
-    val content: String = "",
-    val color: Int = Note.colors.random(),
-    val isLoading: Boolean = false
-)
