--- app/src/main/java/com/manujain/flashnotes/domain/usecase/NotesUsecase.kt
@@ -1,8 +1,0 @@
-package com.manujain.flashnotes.domain.usecase
-
-data class NotesUsecase(
-    val addNote: AddNote,
-    val getNote: GetNote,
-    val getNotes: GetNotes,
-    val deleteNote: DeleteNote
-)
