--- app/src/main/java/com/manujain/flashnotes/domain/utils/NotesOrder.kt
@@ -1,12 +1,0 @@
-package com.manujain.flashnotes.domain.utils
-
-sealed class NotesOrder(val orderType: OrderType) {
-    class Title(orderType: OrderType) : NotesOrder(orderType)
-    class Date(orderType: OrderType) : NotesOrder(orderType)
-    class Color(orderType: OrderType) : NotesOrder(orderType)
-}
-
-enum class OrderType {
-    ASCENDING,
-    DESCENDING
-}
