+++ app/src/main/java/com/manujain/flashnotes/feature_notes/data/data_source/NoteDao.kt
@@ -1,0 +1,25 @@
+package com.manujain.flashnotes.feature_notes.data.data_source
+
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.manujain.flashnotes.feature_notes.domain.model.Note
+import kotlinx.coroutines.flow.Flow
+
+@Dao
+interface NoteDao {
+
+    @Query("SELECT * FROM note")
+    fun getNotes(): Flow<List<Note>>
+
+    @Query("SELECT * FROM note WHERE id = :noteId")
+    suspend fun getNote(noteId: String): Note
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertNote(note: Note)
+
+    @Delete
+    suspend fun deleteNote(note: Note)
+}
