+++ app/src/main/java/com/manujain/flashnotes/feature_notes/data/repository/NotesRepositoryImpl.kt
@@ -1,0 +1,25 @@
+package com.manujain.flashnotes.feature_notes.data.repository
+
+import com.manujain.flashnotes.feature_notes.data.data_source.NoteDao
+import com.manujain.flashnotes.feature_notes.domain.model.Note
+import com.manujain.flashnotes.feature_notes.domain.repository.NotesRepository
+import kotlinx.coroutines.flow.Flow
+
+class NotesRepositoryImpl(private val dao: NoteDao) : NotesRepository {
+
+    override fun getNotes(): Flow<List<Note>> {
+        return dao.getNotes()
+    }
+
+    override suspend fun getNote(noteId: String): Note {
+        return dao.getNote(noteId)
+    }
+
+    override suspend fun insertNote(note: Note) {
+        return dao.insertNote(note)
+    }
+
+    override suspend fun deleteNote(note: Note) {
+        dao.deleteNote(note)
+    }
+}
