+++ app/src/main/java/com/manujain/flashnotes/feature_notes/domain/model/NoteState.kt
@@ -1,0 +1,8 @@
+package com.manujain.flashnotes.feature_notes.domain.model
+
+data class NoteState(
+    val title: String = "",
+    val content: String = "",
+    val color: Int = Note.colors.random(),
+    val isLoading: Boolean = false
+)
