+++ app/src/main/java/com/manujain/flashnotes/feature_notes/domain/repository/NotesRepository.kt
@@ -1,0 +1,15 @@
+package com.manujain.flashnotes.feature_notes.domain.repository
+
+import com.manujain.flashnotes.feature_notes.domain.model.Note
+import kotlinx.coroutines.flow.Flow
+
+interface NotesRepository {
+
+    fun getNotes(): Flow<List<Note>>
+
+    suspend fun getNote(noteId: String): Note
+
+    suspend fun insertNote(note: Note)
+
+    suspend fun deleteNote(note: Note)
+}
