--- app/src/main/java/com/manujain/flashnotes/presentation/notes/NoteSwipeHandler.kt
@@ -1,53 +1,0 @@
-package com.manujain.flashnotes.presentation.notes
-
-import android.graphics.Canvas
-import android.graphics.Color
-import android.graphics.drawable.ColorDrawable
-import androidx.recyclerview.widget.ItemTouchHelper
-import androidx.recyclerview.widget.RecyclerView
-
-/*
-* Handler to support note swipe actions
-* Currently supports logic for RIGHT swipe only
-* Ref: https://github.com/kitek/android-rv-swipe-delete/blob/master/app/src/main/java/pl/kitek/rvswipetodelete/SwipeToDeleteCallback.kt
-*/
-class NoteSwipeHandler(
-    private val onSwiped: (Int) -> Unit
-) : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.RIGHT) {
-
-    private val swipeBackground = ColorDrawable(Color.RED)
-
-    override fun onMove(
-        recyclerView: RecyclerView,
-        viewHolder: RecyclerView.ViewHolder,
-        target: RecyclerView.ViewHolder
-    ): Boolean {
-        return false // INFO: This action is not supported yet
-    }
-
-    override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-        onSwiped(viewHolder.absoluteAdapterPosition)
-    }
-
-    override fun onChildDraw(
-        c: Canvas,
-        recyclerView: RecyclerView,
-        viewHolder: RecyclerView.ViewHolder,
-        dX: Float,
-        dY: Float,
-        actionState: Int,
-        isCurrentlyActive: Boolean
-    ) {
-        val itemView = viewHolder.itemView
-
-        swipeBackground.apply {
-            setBounds(
-                itemView.left,
-                itemView.top,
-                itemView.left + dX.toInt(),
-                itemView.bottom
-            ) // INFO: Handling background for right swipe only
-            draw(c)
-        }
-    }
-}
