+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.manujain.flashnotes.presentation.notes.HomeFragment"
+        android:name="com.manujain.flashnotes.feature_notes.presentation.notes.HomeFragment"
         tools:layout="@layout/fragment_home">
         <action
             android:id="@+id/action_homeFragment_to_addEditNoteFragment"
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/addEditNoteFragment"
-        android:name="com.manujain.flashnotes.presentation.addeditnotes.AddEditNoteFragment"
+        android:name="com.manujain.flashnotes.feature_notes.presentation.addeditnotes.AddEditNoteFragment"
         tools:layout="@layout/fragment_add_edit_note">
     </fragment>
